Feat: Added sequential for forward and backward

+++ nunet/layer/activation/relu.py
@@ -1,5 +1,6 @@
-import numpy as np 
+import numpy as np
 from nunet.layer import Module
+
 
 class ReLU(Module):
     def __init__(self):
@@ -11,10 +12,10 @@ class ReLU(Module):
         self._last_input = input_array
         output = np. maximum(0, input_array)
         return output
-    
+
     def backward(self, gradwrtoutput):
         gradwrtoutput_array = np.array(gradwrtoutput)
-        relu_derivative_mask = (self._last_input > 0).astype(gradwrtoutput_array.dtype)
+        relu_derivative_mask = (self._last_input > 0).astype(
+            gradwrtoutput_array.dtype)
         grad_wrt_input = gradwrtoutput_array * relu_derivative_mask
         return grad_wrt_input
-
